Новый компонент с сортировкой
