+++ project/src/components/offersList/offersList.tsx
@@ -1,7 +1,8 @@
 import {CityCard} from '../cityCard/cityCard';
-import React, {useState} from 'react';
+import React from 'react';
 import {Map} from '../map/map';
 import {Offers} from '../../types/offer';
+import {SortingForm} from '../sortingFom/sortingForm';
 
 
 type CityListProps = {
@@ -11,11 +12,6 @@ type CityListProps = {
 
 export function OffersList({city, offers}: CityListProps) {
 
-  const [openDropDown, setOpenDropDown] = useState(false);
-
-  const dropDownHandle = () => {
-    setOpenDropDown(!openDropDown);
-  };
   const currentOffers = offers.filter((offer) => offer.city.name === city);
   const generateOffersList = currentOffers.map((offer) => <CityCard key={offer.id} offer={offer}/>);
 
@@ -26,28 +22,7 @@ export function OffersList({city, offers}: CityListProps) {
         <section className="cities__places places">
           <h2 className="visually-hidden">Places</h2>
           <b className="places__found">{currentOffers.length} places to stay in {city}</b>
-          <form className="places__sorting" action="#" method="get">
-            <span className="places__sorting-caption">Sort by</span>
-            <span
-              className="places__sorting-type" tabIndex={0}
-              onClick={dropDownHandle}
-            >
-              Popular
-              <svg className="places__sorting-arrow" width="7" height="4">
-                <use xlinkHref="#icon-arrow-select"></use>
-              </svg>
-            </span>
-            <ul
-              className={openDropDown
-                ? 'places__options places__options--custom places__options--opened'
-                : 'places__options places__options--custom'}
-            >
-              <li className="places__option places__option--active" tabIndex={0}>Popular</li>
-              <li className="places__option" tabIndex={0}>Price: low to high</li>
-              <li className="places__option" tabIndex={0}>Price: high to low</li>
-              <li className="places__option" tabIndex={0}>Top rated first</li>
-            </ul>
-          </form>
+          <SortingForm/>
           <div className="cities__places-list places__list tabs__content">
             {generateOffersList}
           </div>
